+++ packages/tests/decorators/withBreakpointManager.spec.ts
@@ -1,10 +1,11 @@
-import { describe, it, expect, jest, beforeEach, afterEach } from 'bun:test';
+import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import {
   Base,
   getInstanceFromElement,
   withBreakpointManager,
   withName,
 } from '@studiometa/js-toolkit';
+import type { Base as BaseType } from '@studiometa/js-toolkit';
 import {
   useMatchMedia,
   resizeWindow,
@@ -30,8 +31,8 @@ async function setupTest() {
       </div>
     </div>
   `;
-  const fn = jest.fn();
-  const withMock = (BaseClass: typeof Base, name) =>
+  const fn = vi.fn();
+  const withMock = (BaseClass: typeof BaseType, name) =>
     withName(
       class extends BaseClass {
         mounted() {
@@ -108,10 +109,10 @@ describe('The withBreakpointManager decorator', () => {
     expect(() => {
       // eslint-disable-next-line no-unused-vars
       class Bar extends withBreakpointManager(withName(Base, 'Bar'), {}) {}
-    }).toThrow(/must be an array/);
+    }).toThrowError(/must be an array/);
     expect(() => {
       // eslint-disable-next-line no-unused-vars
       class Bar extends withBreakpointManager(withName(Base, 'Bar'), [[]]) {}
-    }).toThrow(/at least 2/);
+    }).toThrowError(/at least 2/);
   });
 });