+++ packages/tests/utils/css/getOffsetSizes.spec.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect } from 'bun:test';
+import { describe, it, expect } from 'vitest';
 import { getOffsetSizes } from '@studiometa/js-toolkit/utils';
 
 describe('The `getOffsetSizes` method', () => {